Algos refactor
