+++ Graph_Algorithms/a_start_dijkstra/Nodes.py
@@ -1,0 +1,13 @@
+
+class Nodes:
+    def __init__(self, x, y, row, col, node_id):
+        self.x, self.y = x, y  # row, col
+        self.row, self.col = row, col  # multiple of 5, distance on diangonals will be 7 and horizontal as well as vertical would be 5
+        self.G, self.H = 0, 0  # G - distance from current node to start node , H - heuristic distance from current node to end node
+        self.F = self.G + self.H  # F (node in nodes) = G (node in nodes) + H (node in nodes)
+        self.parent_ptr = 0
+        self.node_id = node_id
+        self.neighbours_lst = []
+
+    def __str__(self):
+        return f"Node at ({self.x}, {self.y}) - Row: {self.row}, Col: {self.col}"
