+++ Graph_Algorithms/a_start_dijkstra/TODOs
@@ -1,0 +1,2 @@
+# TODO: change list for numpy arrays
+# TODO: reafctor code for CUDA GPU
