+++ Graph_Algorithms/ant_colony/UiUtils.py
@@ -1,0 +1,43 @@
+import pygame
+
+from Graph_Algorithms.ConstVars import WIDTH, HEIGHT, WHITE, BLACK, ADJUST_VECTOR, NODE_SIZE, BLUE, YELLOW, GREEN, RED, INNER_HEIGHT, \
+    INNER_WIDTH, ORANGE
+
+def ui_runner(start_pt, goal_pt, grid, obstacles, room_coords, path):
+    screen = pygame.display.set_mode((WIDTH, HEIGHT))
+
+    running = True
+    while running:
+        for event in pygame.event.get():
+            if event.type == pygame.QUIT:
+                running = False
+            elif event.type == pygame.KEYDOWN:
+                if event.key == pygame.K_ESCAPE:
+                    running = False
+
+        screen.fill(BLACK)
+
+        for node in grid:
+            pygame.draw.circle(screen, WHITE, (node.x + ADJUST_VECTOR, node.y + ADJUST_VECTOR), NODE_SIZE)
+
+        for obstacle in obstacles:
+            pygame.draw.rect(screen, BLUE,
+                             (obstacle[0] + ADJUST_VECTOR, obstacle[1] + ADJUST_VECTOR, obstacle[2], obstacle[3]))
+
+        pygame.draw.circle(screen, YELLOW, (start_pt[0] + ADJUST_VECTOR, start_pt[1] + ADJUST_VECTOR), NODE_SIZE * 8)
+        pygame.draw.circle(screen, GREEN, (goal_pt[0] + ADJUST_VECTOR, goal_pt[1] + ADJUST_VECTOR), NODE_SIZE * 8)
+
+        pygame.draw.circle(screen, RED, (INNER_WIDTH + ADJUST_VECTOR, INNER_HEIGHT + ADJUST_VECTOR), NODE_SIZE * 10)
+
+        if path:
+            adjusted_path = [(x + ADJUST_VECTOR, y + ADJUST_VECTOR) for x, y in path]
+            for i in range(1, len(adjusted_path)):
+                pygame.draw.line(screen, RED, adjusted_path[i - 1], adjusted_path[i], NODE_SIZE * 4)
+
+        if len(room_coords) > 1:
+            pygame.draw.lines(screen, ORANGE, True,
+                              [(x + ADJUST_VECTOR, y + ADJUST_VECTOR) for x, y in room_coords], NODE_SIZE)
+
+        pygame.display.update()
+
+    pygame.quit()
