+++ Graph_Algorithms/rrt_rrt_star/EuclideanDistanceUtils.py
@@ -1,0 +1,4 @@
+import math
+
+def euclidean_distance(node1, node2):
+    return math.sqrt((node1.x - node2.x) ** 2 + (node1.y - node2.y) ** 2)
