+++ Graph_Algorithms/rrt_rrt_star/LocalConstVars.py
@@ -1,0 +1,2 @@
+MAX_ITERATIONS = 10000
+EXPANSION_DISTANCE = 20
