+++ Graph_Algorithms/rrt_rrt_star/RRTNodes.py
@@ -1,0 +1,6 @@
+class RRTNode:
+    def __init__(self, x, y):
+        self.x = x
+        self.y = y
+        self.parent = None
+        self.cost = 0
